Use one parent and cramjam-lzo
