--- cramjam-core/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "cramjam-core"
-version = "2.2.0"
-authors = ["milesgranger <miles59923@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-pyo3 = { version = "0.13.2", default-features = false, features = ["macros"] }
-numpy = "0.13.0"
